refactor: assign shared status owners
