+++ packages/shared/src/types/sandbox-events.ts
@@ -1,8 +1,28 @@
 import { z } from "zod";
 import { recordSchema } from "./artifacts";
 import { sessionDiffBaselineRepositorySchema } from "./session-diffs";
-import { gitSyncStatusSchema, type EventType } from "./statuses";
 import { resolvedSessionAttachmentsSchema } from "./session-attachments";
+
+export type GitSyncStatus = "pending" | "in_progress" | "completed" | "failed";
+
+export const gitSyncStatusSchema = z.enum(["pending", "in_progress", "completed", "failed"]);
+
+export type EventType =
+  | "heartbeat"
+  | "ready"
+  | "token"
+  | "tool_call"
+  | "step_start"
+  | "step_finish"
+  | "tool_result"
+  | "git_sync"
+  | "error"
+  | "execution_complete"
+  | "artifact"
+  | "push_complete"
+  | "push_error"
+  | "warning"
+  | "user_message";
 
 export interface AgentEvent {
   id: string;
